backend/tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx>=0.27.2
python-dotenv>=1.0.1 